Rename interactive backend scripts
