+++ pympipool/shared/executorbase.py
@@ -417,20 +417,20 @@ def _get_backend_path(cores: int):
     Get command to call backend as a list of two strings
 
     Args:
-        cores (int): Number of cores used to execute the task, if it is greater than one use mpiexec.py else serial.py
+        cores (int): Number of cores used to execute the task, if it is greater than one use mpiexec_interactive.py else serial_interactive.py
 
     Returns:
         list[str]: List of strings containing the python executable path and the backend script to execute
     """
     command_lst = [sys.executable]
     if cores > 1 and importlib.util.find_spec("mpi4py") is not None:
-        command_lst += [_get_command_path(executable="mpiexec.py")]
+        command_lst += [_get_command_path(executable="mpiexec_interactive.py")]
     elif cores > 1:
         raise ImportError(
             "mpi4py is required for parallel calculations. Please install mpi4py."
         )
     else:
-        command_lst += [_get_command_path(executable="serial.py")]
+        command_lst += [_get_command_path(executable="serial_interactive.py")]
     return command_lst
 
 
@@ -439,7 +439,7 @@ def _get_command_path(executable: str):
     Get path of the backend executable script
 
     Args:
-        executable (str): Name of the backend executable script, either mpiexec.py or serial.py
+        executable (str): Name of the backend executable script, either mpiexec_interactive.py or serial_interactive.py
 
     Returns:
         str: absolute path to the executable script
